requirements.txt
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1694046349000/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1687772187254/work
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.0
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
docker-pycreds==0.4.0
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
fonttools==4.43.0
gitdb==4.0.10
GitPython==3.1.34
Hydra==2.5
hydra-core==1.3.2
idna==3.4
imagecodecs==2023.3.16
imageio==2.31.5
importlib-resources==6.0.1
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1655241626755/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1653754926575/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1690896916983/work
joblib==1.3.2
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1633454794268/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1669775081186/work
kiwisolver==1.4.5
lazy_loader==0.3
matplotlib==3.7.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
MedPy==0.4.0
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
networkx==3.1
nibabel==5.1.0
numpy==1.24.4
omegaconf==2.3.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==2.0.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pathtools==0.1.2
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==10.0.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1688565951714/work
protobuf==4.24.2
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1691408637400/work
pyparsing==3.1.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==19.0.2
qudida==0.0.4
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==1.30.0
setproctitle==1.3.2
SimpleITK==2.3.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.0
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
threadpoolctl==3.2.0
tifffile==2023.7.10
torch==1.7.1
torchaudio==0.7.2
torchsummary==1.5.1
torchvision==0.8.2
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827257044/work
tqdm==4.66.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wandb==0.15.9
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
zipp==3.16.2
